LPS-80967 Portlet 3.0: Decouple request/response factory APIs from implementation details (moved PortletSessionImpl)
